Created base mythic html template and cleaned up games_list html.
Also changed the way games are sent to the list - now the existing ones are sent in the template.
Changed the game list view - now it handles adding to games to db and deleting games. Also changed what it sends after posting a new game - now it also sends an url for that game